# relocate parser look ahead loops