Part of #1823.

This PR relocates the read ahead loops into the proper switch cases. It further reduces code complexity (less branching) with a slight speedup for the optimized actions.