@jerch 2019-06-05
Speed comparison (tested with [this script](https://gist.github.com/jerch/827f18ba4af0b214ff45079403f5a1a3)):

- before:
```
   Context "examples/parser_throughput.js"
      Context "Parser performance - 50MB data"
         Context "print - a"
            Case "throughput" : 5 runs - average throughput: 432.44 MB/s
         Context "execute - \n"
            Case "throughput" : 5 runs - average throughput: 95.12 MB/s
         Context "escape - ESC E"
            Case "throughput" : 5 runs - average throughput: 53.92 MB/s
         Context "escape with collect - ESC % G"
            Case "throughput" : 5 runs - average throughput: 22.37 MB/s
         Context "simple csi - CSI A"
            Case "throughput" : 5 runs - average throughput: 84.30 MB/s
         Context "csi with collect - CSI ? p"
            Case "throughput" : 5 runs - average throughput: 55.03 MB/s
         Context "csi with params - CSI 1;2 m"
            Case "throughput" : 5 runs - average throughput: 104.91 MB/s
         Context "osc (small payload) - OSC 0;hi ST"
            Case "throughput" : 5 runs - average throughput: 50.15 MB/s
         Context "osc (big payload) - OSC 0;<text> ST"
            Case "throughput" : 5 runs - average throughput: 48.93 MB/s
         Context "DCS (small payload)"
            Case "throughput" : 5 runs - average throughput: 49.25 MB/s
         Context "DCS (big payload)"
            Case "throughput" : 5 runs - average throughput: 94.54 MB/s
```
- after:
```
   Context "examples/parser_throughput.js"
      Context "Parser performance - 50MB data"
         Context "print - a"
            Case "throughput" : 5 runs - average throughput: 632.97 MB/s
         Context "execute - \n"
            Case "throughput" : 5 runs - average throughput: 89.43 MB/s
         Context "escape - ESC E"
            Case "throughput" : 5 runs - average throughput: 58.36 MB/s
         Context "escape with collect - ESC % G"
            Case "throughput" : 5 runs - average throughput: 22.37 MB/s
         Context "simple csi - CSI A"
            Case "throughput" : 5 runs - average throughput: 89.36 MB/s
         Context "csi with collect - CSI ? p"
            Case "throughput" : 5 runs - average throughput: 94.44 MB/s
         Context "csi with params - CSI 1;2 m"
            Case "throughput" : 5 runs - average throughput: 124.12 MB/s
         Context "osc (small payload) - OSC 0;hi ST"
            Case "throughput" : 5 runs - average throughput: 50.08 MB/s
         Context "osc (big payload) - OSC 0;<text> ST"
            Case "throughput" : 5 runs - average throughput: 48.31 MB/s
         Context "DCS (small payload)"
            Case "throughput" : 5 runs - average throughput: 60.97 MB/s
         Context "DCS (big payload)"
            Case "throughput" : 5 runs - average throughput: 256.35 MB/s
```